Set the line spacing directly
When passing it as an option to the setspace package, in some cases not
the last option would set the effective line spacing since the options
are not processed in calling order.

By calling the `\singlespacing,...` commands at the end of the package,
the last called option determines the line spacing.